added jwt stuff

+++ resources/item.py
@@ -1,8 +1,7 @@
-from ast import Try
-import uuid
-from flask import request
 from flask.views import MethodView
 from flask_smorest import abort, Blueprint
+from flask_jwt_extended import jwt_required, get_jwt
+
 from models.item import ItemModel
 from schemas import ItemSchema, ItemUpdateSchema
 from sqlalchemy.exc import SQLAlchemyError
@@ -12,10 +11,12 @@ blp = Blueprint("Items", __name__, description="Operations on items.")
 
 @blp.route("/item")
 class ItemList(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema(many=True))
     def get(self):
         return ItemModel.query.all()
     
+    @jwt_required(fresh=True)
     @blp.arguments(ItemSchema)
     @blp.response(201, ItemSchema)
     def post(self, item_data):
@@ -28,13 +29,14 @@ class ItemList(MethodView):
         
         return item
 
-@blp.route("/item/<string:item_id>")
+@blp.route("/item/<int:item_id>")
 class Item(MethodView):
     @blp.response(200, ItemSchema)
     def get(self, item_id):
         item = ItemModel.query.get_or_404(item_id)
         return item
     
+    @jwt_required()
     @blp.arguments(ItemUpdateSchema)
     @blp.response(200, ItemSchema)
     def put(self, item_data, item_id):
@@ -50,8 +52,11 @@ class Item(MethodView):
         
         return item
         
-    
+    @jwt_required()
     def delete(self, item_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
